compilation failure on windows
libmariadb\libmariadb\ma_stmt_codec.c(263): warning C4244: '*=': conversion from 'double' to 'unsigned long', possible loss of data

generally avoid slow floating point pow() when all you need is to
multiply an integer by a small power of 10.

+++ unittest/libmariadb/ps_bugs.c
@@ -4821,7 +4821,7 @@ static int test_conc_fraction(MYSQL *mysql)
 
     diag("second_part: %ld", tm.second_part);
 
-    expected= 9 * powl(10, (5 - i));
+    expected= 9 * (unsigned int)powl(10, (5 - i));
 
     FAIL_IF(tm.second_part != expected, "expected fractional part to be 900000");
